place the check wrapper in a single place
