+++ skrf/calibration/deembedding.py
@@ -61,30 +61,11 @@ import numpy as np
 from numpy import concatenate, conj, flip, real, angle, exp, zeros
 from numpy.fft import fft, fftshift, irfft, ifftshift
 from scipy.interpolate import interp1d
-from functools import wraps
 try:
+    from . import plotting    # will perform the correct setup for matplotlib before it is called below
     import matplotlib.pyplot as plt
 except ImportError:
     plt = None
-    
-def check_plotting(func):
-    """
-    This decorator checks if matplotlib.pyplot is available under the name mplt.
-    If not, raise an RuntimeError.
-
-    Raises
-    ------
-    RuntimeError
-        When trying to run the decorated function without matplotlib
-    """
-
-    @wraps(func)
-    def wrapper(*args, **kwargs):
-        if plt is None:
-            raise RuntimeError('Plotting is not available')
-        func(*args, **kwargs)
-
-    return wrapper
 
 
 class Deembedding(ABC):
