+++ skrf/vectorFitting.py
@@ -19,26 +19,6 @@ import warnings
 from timeit import default_timer as timer
 
 
-def check_plotting(func):
-    """
-    This decorator checks if matplotlib.pyplot is available under the name mplt.
-    If not, raise an RuntimeError.
-
-    Raises
-    ------
-    RuntimeError
-        When trying to run the decorated function without matplotlib
-    """
-
-    @wraps(func)
-    def wrapper(*args, **kwargs):
-        if mplt is None:
-            raise RuntimeError('Plotting is not available')
-        func(*args, **kwargs)
-
-    return wrapper
-
-
 class VectorFitting:
     """
     This class provides a Python implementation of the Vector Fitting algorithm and various functions for the fit
